implemented option to use modes with delimiters
