+++ signaturenDruck/js/classes/ShelfmarksFromSRUData.js
@@ -3,6 +3,8 @@ const _ = require('lodash')
 const Shelfmark = require('../shelfmark.js')
 const Modes = require('./Modes.js')
 const config = remote.getGlobal('config')
+const Formats = require('../classes/Formats')
+const FormatLinesByMode = require('../classes/FormatLinesByMode')
 
 class ShelfmarksFromSRUData {
   /*
@@ -29,6 +31,8 @@ class ShelfmarksFromSRUData {
   getShelfmark (xml) {
     let sig = new Shelfmark()
     let mode = new Modes()
+    let formats = new Formats()
+    let formatArray = formats.formats
 
     sig.error = getError(xml)
     if (sig.error === '') {
@@ -46,17 +50,24 @@ class ShelfmarksFromSRUData {
           'lines': ''
         }
         data.name = value.format
-        let regex = new RegExp(value.regEx)
-        if (regex.test(sig.txtOneLine) && sig.defaultSubMode === '') {
-          sig.defaultSubMode = value.id
-        }
-        let lines = sig.txtOneLine.match(regex)
-        if (lines !== null) {
-          lines.shift()
-        }
-        data.lines = lines
-        if (data.lines !== null) {
-          data.lines = formatData(sig.location, data.lines, value.result)
+        if (value.useRegEx) {
+          let regex = new RegExp(value.regEx)
+          if (regex.test(sig.txtOneLine) && sig.defaultSubMode === '') {
+            sig.defaultSubMode = value.id
+          }
+          let lines = sig.txtOneLine.match(regex)
+          if (lines !== null) {
+            lines.shift()
+          }
+          data.lines = lines
+          if (data.lines !== null) {
+            data.lines = FormatLinesByMode.formatLines(sig.location, data.lines, value.result)
+          }
+        } else {
+          data.lines = sig.txtOneLine.split(value.delimiter)
+          if (data.lines !== null) {
+            data.lines = FormatLinesByMode.formatLines(sig.location, data.lines, value.result, formatArray[value.format].lines)
+          }
         }
         sig.subModes.push(data)
       })
@@ -100,28 +111,6 @@ function getError (object) {
   } else {
     return 'Barcode wurde nicht gefunden'
   }
-}
-
-function formatData (location, lines, linesOrdered) {
-  let data = []
-  let i = 0
-  _.each(linesOrdered, function (value) {
-    let regex1 = new RegExp(/\$\d{1,3}/)
-    let regex2 = new RegExp(/\$LOC/)
-    while (regex1.test(value)) {
-      let matches = value.match(regex1)
-      _.each(matches, function (match) {
-        let group = match.split('$')[1]
-        value = value.replace(match, lines[group - 1])
-      })
-    }
-    if (regex2.test(value)) {
-      value = value.replace('$LOC', location)
-    }
-    data[i] = value
-    i++
-  })
-  return data
 }
 
 module.exports = ShelfmarksFromSRUData
