+++ signaturenDruck/js/preK10plus/loadDataFromFileOld.js
@@ -7,6 +7,8 @@ const DataExtractOld = require('./dataExtractOld.js')
 // requires the shelfmark class
 const Shelfmark = require('../shelfmark.js')
 const Modes = require('../classes/Modes')
+const Formats = require('../classes/Formats')
+const FormatLinesByMode = require('../classes/FormatLinesByMode')
 
 module.exports = function (allLines) {
   let obj = {
@@ -17,6 +19,8 @@ module.exports = function (allLines) {
   let ppnAktuell = ''
   let plainTxt = ''
   let mode = new Modes()
+  let formats = new Formats()
+  let formatArray = formats.formats
 
   allLines.map((line) => {
     let first4 = extract.firstFour(line)
@@ -31,22 +35,29 @@ module.exports = function (allLines) {
       sig.txtOneLine = plainTxt
       let allSubModeData = mode.modes[config.get('mode.defaultMode')].subModes
       _.forEach(allSubModeData, function (value) {
-        let regex = new RegExp(value.regEx)
         let data = {
           'name': '',
           'lines': ''
         }
         data.name = value.format
-        if (regex.test(plainTxt) && sig.defaultSubMode === '') {
-          sig.defaultSubMode = value.id
-        }
-        let lines = plainTxt.match(regex)
-        if (lines !== null) {
-          lines.shift()
-        }
-        data.lines = lines
-        if (data.lines !== null) {
-          data.lines = formatData(sig.location, data.lines, value.result)
+        if (value.useRegEx) {
+          let regex = new RegExp(value.regEx)
+          if (regex.test(plainTxt) && sig.defaultSubMode === '') {
+            sig.defaultSubMode = value.id
+          }
+          let lines = plainTxt.match(regex)
+          if (lines !== null) {
+            lines.shift()
+          }
+          data.lines = lines
+          if (data.lines !== null) {
+            data.lines = FormatLinesByMode.formatLines(sig.location, data.lines, value.result)
+          }
+        } else {
+          data.lines = plainTxt.split(value.delimiter)
+          if (data.lines !== null) {
+            data.lines = FormatLinesByMode.formatLines(sig.location, data.lines, value.result, formatArray[value.format].lines)
+          }
         }
         sig.subModes.push(data)
       })
@@ -60,26 +71,4 @@ module.exports = function (allLines) {
     }
   })
   return obj
-}
-
-function formatData (location, lines, linesOrdered) {
-  let data = []
-  let i = 0
-  _.each(linesOrdered, function (value) {
-    let regex1 = new RegExp(/\$\d{1,3}/)
-    let regex2 = new RegExp(/\$LOC/)
-    while (regex1.test(value)) {
-      let matches = value.match(regex1)
-      _.each(matches, function (match) {
-        let group = match.split('$')[1]
-        value = value.replace(match, lines[group - 1])
-      })
-    }
-    if (regex2.test(value)) {
-      value = value.replace('$LOC', location)
-    }
-    data[i] = value
-    i++
-  })
-  return data
 }
